config: use uint32 for validators/committee members count

+++ pkg/rpcclient/actor/maker.go
@@ -214,7 +214,7 @@ func (a *Actor) CalculateValidUntilBlock() (uint32, error) {
 	if err != nil {
 		return 0, fmt.Errorf("can't get block count: %w", err)
 	}
-	var vc = int(a.version.Protocol.ValidatorsCount)
+	var vc = uint32(a.version.Protocol.ValidatorsCount)
 	var bestH = uint32(0)
 	for h, n := range a.version.Protocol.ValidatorsHistory { // In case it's enabled.
 		if h >= bestH && h <= blockCount {
@@ -223,5 +223,5 @@ func (a *Actor) CalculateValidUntilBlock() (uint32, error) {
 		}
 	}
 
-	return blockCount + uint32(vc+1), nil
+	return blockCount + vc + 1, nil
 }
